movie_rec/requirements.txt
cachelib==0.4.1
certifi==2021.10.8
click==8.0.3
Flask==2.0.2
Flask-Session==0.4.0
gunicorn==20.1.0
import-ipynb==0.1.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
numpy==1.21.4
pandas==1.3.5
python-dateutil==2.8.2
pytz==2021.3
scikit-learn==1.0.1
scipy==1.7.3
six==1.16.0
sklearn-features==0.0.2
threadpoolctl==3.0.0
Werkzeug==2.0.2
